requirements.txt
psycopg2-binary
faker
pymongo
cassandra-driver
cython
cryptography
oracledb #python -m pip install oracledb --no-deps
